Fix for texture mapping at apex (north pole) of sphere
Shared vertices were used to achieve smoothing (normals emulating smoothness) but at north pole the sharing of the apex vertex's texture coordinates was causing a swirling effect.  Adding the same point as a vertex multiple times but with different texture coordinates causes the smoothing feature to stop working.  The fix was to apply a "fudge factor" to the y coordinate at the apex, so it is a minutely different point for each of the top triangles.  This allows the texture (UV) mapping to work properly as well as the smoothing.